backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
openai>=1.3.7
aiohttp>=3.9.1
Pillow>=10.4.0
reportlab>=4.0.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.1
celery>=5.3.4
bcrypt>=4.0.1

# PostgreSQL数据库驱动
psycopg2-binary>=2.9.7
# 异步PostgreSQL支持（可选，用于async/await）
asyncpg>=0.29.0

# 数据库迁移工具
alembic>=1.12.0
